refacto composant fonctionnel
