+++ kasa-app/src/components/host/Host.module.scss
@@ -1,0 +1,28 @@
+@import "../../styles/index";
+
+.owner {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  @include media-query(desktop) {
+    gap: 1rem;
+  }
+  & img {
+    width: 32px;
+    height: 32px;
+    border-radius: 100%;
+    object-fit: contain;
+    @include media-query(desktop) {
+      width: 64px;
+      height: 64px;
+    }
+  }
+  & p {
+    font-size: 12px;
+    color: $primary-bg-kasa-color;
+    font-weight: 500;
+    @include media-query(desktop) {
+      font-size: 18px;
+    }
+  }
+}
