+++ kasa-app/src/components/rating/Rating.module.scss
@@ -1,0 +1,15 @@
+@import "../../styles/index";
+
+.rates {
+  display: flex;
+  gap: 0.2rem;
+  color: $primary-bg-kasa-color;
+  font-size: 18px;
+  @include media-query(desktop) {
+    font-size: 24px;
+  }
+  & svg:hover {
+    cursor: pointer;
+    color: $primary-bg-kasa-color;
+  }
+}
