+++ kasa-app/src/views/logement/Logement.module.scss
@@ -6,67 +6,6 @@
   gap: 1rem;
 }
 
-.introLogementCarroussel {
-  display: flex;
-  max-width: 100%;
-  height: 225px;
-  overflow: hidden;
-  object-fit: contain;
-  position: relative;
-  @include media-query(desktop) {
-    height: 415px;
-  }
-}
-
-.carrouselItem {
-  border-radius: 1rem;
-  width: 100%;
-  object-fit: cover;
-  object-position: center;
-}
-
-.CarrousselCounter {
-  position: absolute;
-  bottom: 10px;
-  left: 50%;
-  transform: translateX(-50%);
-  color: white;
-  font-size: 14px;
-  padding: 0.2rem 0.6rem;
-  border-radius: 0.5rem;
-}
-
-.buttons {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  & button {
-    background: none;
-    color: white;
-    border: none;
-    cursor: pointer;
-  }
-
-  .left {
-    position: absolute;
-    left: 10px;
-    transform: translateY(-50%);
-  }
-
-  .right {
-    position: absolute;
-    right: 10px;
-    transform: translateY(-50%);
-  }
-
-  .hidden {
-    display: none;
-  }
-}
-
 .infosContainer {
   display: flex;
   flex-direction: column;
@@ -107,46 +46,5 @@
   @include media-query(desktop) {
     flex-direction: column-reverse;
     gap: 1rem;
-  }
-}
-
-.rates {
-  display: flex;
-  gap: 0.2rem;
-  color: $primary-bg-kasa-color;
-  font-size: 18px;
-  @include media-query(desktop) {
-    font-size: 24px;
-  }
-  & svg:hover {
-    cursor: pointer;
-    color: $primary-bg-kasa-color;
-  }
-}
-
-.owner {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  @include media-query(desktop) {
-    gap: 1rem;
-  }
-  & img {
-    width: 32px;
-    height: 32px;
-    border-radius: 100%;
-    object-fit: contain;
-    @include media-query(desktop) {
-      width: 64px;
-      height: 64px;
-    }
-  }
-  & p {
-    font-size: 12px;
-    color: $primary-bg-kasa-color;
-    font-weight: 500;
-    @include media-query(desktop) {
-      font-size: 18px;
-    }
   }
 }
